+++ QhenkiX/qhenki/graphics/d3d12/d3d12_descriptor_heap.h
@@ -19,7 +19,7 @@ class D3D12DescriptorHeap
     std::mutex m_mutex;
 
     size_t m_pointer = 0;
-    std::vector<UINT64> m_free_list;
+    std::vector<size_t> m_free_list;
 
 public:
     const D3D12_DESCRIPTOR_HEAP_DESC& desc = m_desc;
@@ -27,8 +27,8 @@ public:
     bool create(ID3D12Device* device, const D3D12_DESCRIPTOR_HEAP_DESC& desc);
 
     // Thread safe
-    bool allocate(UINT64* alloc_offset);
-    void deallocate(UINT64* alloc_offset);
+    bool allocate(size_t* alloc_offset);
+    void deallocate(size_t* alloc_offset);
 
     /**
      * Converts count number of descriptors into size in bytes
@@ -37,11 +37,12 @@ public:
      */
     unsigned descriptor_count_to_bytes(unsigned count) const;
 
-    // offsets are additional to start of heap
+    // Offsets are additional to start of heap
     void get_CPU_descriptor(D3D12_CPU_DESCRIPTOR_HANDLE* handle,
                             size_t offset_bytes,
                             size_t num_descriptor_offset) const;
-    // offsets are additional to start of heap
+
+    // Offsets are additional to start of heap
     bool get_GPU_descriptor(D3D12_GPU_DESCRIPTOR_HANDLE* handle,
                             size_t offset_bytes,
                             size_t num_descriptor_offset) const;
